Put Set, Variable and Index classes into separate modules. Avoid useless inheritance from BaseItem class (the class has deleted).
